Implement sip.
Sip allows you to read a partial payload. I'm going to use this to
implement sipping headers from specific positions in a file.

In order to implement sip I had to introduce the concept of blocks. We
have parts that split the body into separate buffers, but we don't
checksum the individual parts. This means we can't sip a part and
perform a checksum. We need to perform the checksum on the entire body,
then split the body into parts.

Now we have blocks of parts and the blocks have checksums. When we sip,
we sip blocks which can contain one or more parts.

Closes #32.